Mysql-wordcloud-FTP/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.


matplotlib == 3.3.3


mysql == 0.0.2


numpy == 1.19.5


omelette == 1.1.6


pandas == 1.2.0


pillow == 8.1.0


pillow-simd == 7.0.0.post3


wordcloud == 1.8.1
